requirements.txt
# FPT Shop Contract Checker - Backend Dependencies

# Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6

# Database
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23

# Authentication & Security
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
PyJWT==2.8.0

# HTTP Requests
requests==2.31.0
requests[socks]==2.31.0

# Data Processing
brotli==1.1.0
python-dateutil==2.8.2

# WebSocket Support
python-socketio==5.8.0
eventlet==0.33.3

# Utilities
user-agents==2.2.0
fake-useragent==1.4.0

# Development
werkzeug==2.3.7
python-dotenv==1.0.0